split progress parser gui into an internal package in separate appropriate files
